+++ src/onegov.activity/onegov/activity/models/invoice_item.py
@@ -1,3 +1,5 @@
+import hashlib
+
 from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
@@ -7,6 +9,7 @@ from sqlalchemy import ForeignKey
 from sqlalchemy import Numeric
 from sqlalchemy import Text
 from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy_utils import observes
 from uuid import uuid4
 
 
@@ -23,6 +26,9 @@ class InvoiceItem(Base, TimestampMixin):
 
     #: the invoice group (all items with the same text make one invoice)
     invoice = Column(Text, nullable=False)
+
+    #: the code of the invoice used to identify the invoice through e-banking
+    code = Column(Text, nullable=False)
 
     #: the item group (all items with the same text are visually grouped)
     group = Column(Text, nullable=False)
@@ -46,3 +52,17 @@ class InvoiceItem(Base, TimestampMixin):
     @hybrid_property
     def amount(self):
         return self.unit * self.quantity
+
+    @observes('invoice', 'username')
+    def invoice_username_observer(self, invoice, username):
+        # there's no guarantee that this code is unique for an invoice, though
+        # the chance of it overlapping is very very small -> any algorithm
+        # doing some kind of matching has to account for this fact (probably
+        # by throwing an error)
+        #
+        # we can solve this by introducing a separate invoice record in the
+        # future
+        self.code = ''.join((
+            hashlib.sha1((invoice + username).encode('utf-8')).hexdigest()[:5],
+            hashlib.sha1(username.encode('utf-8')).hexdigest()[:5]
+        ))